Release SDK v2.5.0 (#134)
* ENG-3970: Android Integration Tests FOCUS and BLUR events (#129)

* Make sure there is a FOCUS event before all Field inputs. Make sure there is a BLUR event AFTER input is complete.

* updating libraries

* ENG-3956: Github workflow Android SDK
 * Send alert to slack on failure

* Release Notification (#131)

* Release version 2.4.2

* Updating action

* Fixing setUserID test

* Fixing action (#132)

* updating version
