+++ .github/workflows/github-actions-master-push-assamble.yml
@@ -75,3 +75,15 @@ jobs:
           user_name: 'robot'
           destination_branch: 'main'
           commit_message: ${{ github.event.head_commit.message }}
+
+      - name: Send Slack Notification on Failure
+        if: ${{ failure() }}
+        uses: rtCamp/action-slack-notify@v2
+        env:
+          SLACK_CHANNEL: mobile-sdk-tech-team
+          SLACK_COLOR: ${{ job.status }}
+          SLACK_ICON: https://github.com/rtCamp.png?size=48
+          SLACK_MESSAGE: "Failed execution on master sdk build"
+          SLACK_TITLE: Failed master SDK
+          SLACK_USERNAME: rtBot
+          SLACK_WEBHOOK: ${{ secrets.SLACK_WEBHOOK }}
