+++ .github/workflows/github-actions-release-notification.yml
@@ -1,0 +1,22 @@
+name: After Release SDK Version
+
+on:
+  release:
+    types: [published]
+
+jobs:
+  after-release:
+    runs-on: ubuntu-latest
+
+    steps:
+      - name: "Slack Notification"
+        uses: slackapi/slack-github-action@v1.21.0
+        with:
+          payload: |
+            {
+              "username": "NeuroID",
+              "icon_emoji": ":white_check_mark:",
+              "text": "Android SDK Version: ${{github.event.release.name}}."
+            }
+        env:
+          SLACK_WEBHOOK_URL: ${{inputs.SLACK_WEBHOOK}}
