+++ app/src/main/AndroidManifest.xml
@@ -16,8 +16,7 @@
         tools:ignore="DataExtractionRules,UnusedAttribute">
         <activity
             android:name=".activities.MainActivity"
-            android:exported="true"
-            android:windowSoftInputMode="stateHidden|adjustResize">
+            android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
