+++ app/src/main/java/com/sample/neuroid/us/fragments/NIDFirstFragment.kt
@@ -2,6 +2,8 @@ package com.sample.neuroid.us.fragments
 
 import android.content.Context
 import android.os.Bundle
+import android.text.Editable
+import android.text.TextWatcher
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -46,6 +48,21 @@ class NIDFirstFragment: Fragment() {
                     println("--------------------- onItemClickListener Example")
                 }
             }
+
+            editTextNormalField.addTextChangedListener(object : TextWatcher {
+                override fun beforeTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
+                    println("----------------- beforeTextChanged")
+                }
+
+                override fun onTextChanged(p0: CharSequence?, p1: Int, p2: Int, p3: Int) {
+                    println("----------------- onTextChanged")
+                }
+
+                override fun afterTextChanged(p0: Editable?) {
+                    println("----------------- afterTextChanged")
+                }
+
+            })
         }
     }
 }
