+++ version.properties
@@ -1,5 +1,5 @@
 #Fri May 27 05:08:14 UTC 2022
-SDK_VERSION=2.4.2
+SDK_VERSION=2.5.0
 VERSION_MAJOR=2
-VERSION_MINOR=4
-VERSION_PATCH=2
+VERSION_MINOR=5
+VERSION_PATCH=0
